add the ability to create k8s services for worker pods
sometimes you want your worker pods to be available from outside the
cluster where they're running (e.g. when connecting containerised and
bare-metal applications)

one way to achieve that in Kubernetes is to create a service that unites
one more more pods under a single *stable* IP and sets up port
forwarding from host machines to pods (see
https://kubernetes.io/docs/concepts/services-networking/service/#type-nodeport
for details)
the IP will be internal and it's the user's responsibility to make it
routable and discoverable from outside the cluster

this patch adds a new config field for NodePort settings
if specified, for every worker pod, a service with those settings will be
created
when a worker pod is stopped, the service gets removed

here's a working example of a NodePort service config:

    c.KubeSpawner.nodeports = [
            {'port': 4040, 'name': 'webui'},
            {'port': 5050, 'name': 'blockmanager'},
            {'port': 5051, 'name': 'control'}
    ]

this definition is passed directly into the service spec and therefore
must conform to the
https://kubernetes.io/docs/api-reference/v1.6/#serviceport-v1-core spec

port settings are be exposed as environment variables and for the
snippet above the following variables will be set:

    JPY_NODEPORT_WEBUI=4040:32578:TCP
    JPY_NODEPORT_BLOCKMANAGER=32777:5050:TCP
    JPY_NODEPORT_CONTROL=5051:30000:TCP

the format is POD_PORT:NODE_PORT:PROTOCOL

NODE_PORT, if not configured explicitly, will be allocated by Kubernetes
from a free random high port

so what this actually means is that applications running outside the
cluster can connect to the service IP on port e.g. 32578 and the
connection will be forwarded to the worker pod on port 4040